ledger_proto

Trait EncDec

pub trait EncDec<'a, E = Error>: Encode<Error = E> + Decode<'a, Output = Self, Error = E> { }

Composite trait requiring an object is reversibly encodable and decodable into borrowed types, useful for simplifying type bounds / generics.

(ie. Self == <Self as Decode>::Output, <Self as Encode::Error> == <Self as Decode::Error>)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.